backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
scipy
Pillow
python-multipart
